+++ geopandas/base.py
@@ -1109,6 +1109,38 @@ GeometryCollection
         """
         return _delegate_geo_method("make_valid", self)
 
+    def reverse(self):
+        """Returns a ``GeoSeries`` with the order of coordinates reversed.
+
+        Examples
+        --------
+
+        >>> from shapely.geometry import Polygon, LineString, Point
+        >>> s = geopandas.GeoSeries(
+        ...     [
+        ...         Polygon([(0, 0), (1, 1), (0, 1)]),
+        ...         LineString([(0, 0), (1, 1), (1, 0)]),
+        ...         Point(0, 0),
+        ...     ]
+        ... )
+        >>> s
+        0    POLYGON ((0.00000 0.00000, 1.00000 1.00000, 0....
+        1    LINESTRING (0.00000 0.00000, 1.00000 1.00000, ...
+        2                              POINT (0.00000 0.00000)
+        dtype: geometry
+
+        >>> s.reverse()
+        0    POLYGON ((0.00000 0.00000, 0.00000 1.00000, 1....
+        1    LINESTRING (1.00000 0.00000, 1.00000 1.00000, ...
+        2    POINT (0.00000 0.00000)
+        dtype: geometry
+
+        See also
+        --------
+        GeoSeries.normalize : normalize order of coordinates
+        """
+        return _delegate_geo_method("reverse", self)
+
     def segmentize(self, max_segment_length):
         """Returns a ``GeoSeries`` with vertices added to line segments based on
         maximum segment length.
@@ -4283,7 +4315,7 @@ GeometryCollection
 
         return pd.Series(distances, index=self.index, name="hilbert_distance")
 
-    def sample_points(self, size, method="uniform", seed=None, **kwargs):
+    def sample_points(self, size, method="uniform", seed=None, rng=None, **kwargs):
         """
         Sample points from each geometry.
 
@@ -4314,8 +4346,8 @@ GeometryCollection
             http://pysal.org/pointpats/api.html#random-distributions). Pointpats methods
             are implemented for (Multi)Polygons only and will return an empty MultiPoint
             for other geometry types.
-        seed : {None, int, array_like[ints], SeedSequence, BitGenerator, Generator}, optional
-            A seed to initialize the numpy BitGenerator. If None, then fresh,
+        rng : {None, int, array_like[ints], SeedSequence, BitGenerator, Generator}, optional
+            A random generator or seed to initialize the numpy BitGenerator. If None, then fresh,
             unpredictable entropy will be pulled from the OS.
         **kwargs : dict
             Options for the pointpats sampling algorithms.
@@ -4343,13 +4375,19 @@ GeometryCollection
         from .geoseries import GeoSeries
         from .tools._random import uniform
 
+        if seed is not None:
+            warn(
+                "The 'seed' keyword is deprecated. Use 'rng' instead.",
+                FutureWarning,
+                stacklevel=2,
+            )
+            rng = seed
+
         if method == "uniform":
             if pd.api.types.is_list_like(size):
-                result = [
-                    uniform(geom, s, seed) for geom, s in zip(self.geometry, size)
-                ]
+                result = [uniform(geom, s, rng) for geom, s in zip(self.geometry, size)]
             else:
-                result = self.geometry.apply(uniform, size=size, seed=seed)
+                result = self.geometry.apply(uniform, size=size, rng=rng)
 
         else:
             pointpats = compat.import_optional_dependency(