+++ CHANGELOG.md
@@ -7,6 +7,10 @@ PyGEOS are installed.  PyGEOS will continue to be used by default when PyGEOS is
 installed alongside Shapely < 2.0.  Support for PyGEOS and Shapely < 2.0 will be removed
 in GeoPandas 1.0.
 
+API changes:
+
+- ``seed`` keyword in ``sample_points`` is deprecated. Use ``rng`` instead. (#2913).
+
 New methods:
 
 - Added ``remove_repeated_points`` method from shapely to GeoSeries/GeoDataframe (#2940).
@@ -19,6 +23,7 @@ New methods:
 - Added ``offset_curve`` method from shapely to GeoSeries/GeoDataframe (#2902).
 - Added ``shortest_line`` method from shapely to GeoSeries/GeoDataframe (#2960).
 - Added ``minimum_rotated_rectangle`` method from shapely to GeoSeries/GeoDataframe (#2541).
+- Added ``reverse`` method from shapely to GeoSeries/GeoDataframe (#2988).
 
 New features and improvements: